Ability to override db module
